+++ src/solc_doc/DanglingBytes.sol
@@ -5,7 +5,7 @@ pragma solidity >=0.8.0 <0.9.0;
 contract C {
     bytes x = "012345678901234567890123456789";
 
-    function test() external returns(uint) {
+    function test() external returns (uint256) {
         (x.push(), x.push()) = (0x01, 0x02);
         return x.length;
     }
